backend/requirements.txt
fastapi==0.85.2
uvicorn==0.19.0
python-multipart==0.0.5
pdf2image==1.16.2
Pillow==9.2.0
PyPDF2==2.6.0
typing_extensions==4.3.0